requirements.txt
flet~=0.19.0
pandas~=2.2.0
opencv-python~=4.9.0.80
PyAutoGUI~=0.9.54
numpy~=1.26.3
pillow~=10.2.0